modified:   README.md deleted:    enable_cpu-perf.sh deleted:    enable_cpu-status.sh modified:   enable_fantastic-packages.sh deleted:    enable_internet-detector.sh deleted:    enable_ipinfo.sh deleted:    enable_log-viewer.sh deleted:    enable_temp-status.sh modified:   enable_ttl.sh

+++ enable_fantastic-packages.sh
@@ -5,8 +5,17 @@ set -x
 
 vNAME="fantastic"
 vBASE_URL="https://${vNAME}-packages.github.io/packages/releases"
+
 vALLOW_VER="("23.05" "24.10" "SNAPSHOT")"
 vALLOW_ARCH="("aarch64_cortex-a53" "aarch64_generic" "arm_cortex-a15_neon-vfpv4" "mips_24kc" "mipsel_24kc" "x86_64")"
+vALLOW_PKGS="luci-app-cpu-perf luci-i18n-cpu-perf-ru
+luci-app-cpu-status luci-i18n-cpu-status-ru
+internet-detector luci-app-internet-detector luci-i18n-internet-detector-ru
+luci-app-ipinfo
+luci-app-log-viewer luci-i18n-log-viewer-ru
+luci-app-temp-status luci-i18n-temp-status-ru
+"
+
 vOWRT_VER="$(cat /etc/banner | grep OpenWrt | sed 's/\.0.*$/ /g' | awk '{print $2}')"
 vARCH="$(opkg print-architecture | tail -n 1 | awk '{print $2}')"
 vWGET_CMD="wget -q --show-progress"
@@ -21,15 +30,31 @@ else
         echo " ### NEED arch -> $vALLOW_ARCH ###"
         echo " ###"
     else
-        echo "ku-ku epta!!!"
-        vKEYRING_FILE="$(curl -s ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/ | grep ${vNAME}-keyring | sed "s/\"/ /g" | awk '{print $5}')"
-        vFEEDS_FILE="$(curl -s ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/ | grep ${vNAME}-packages-feeds | sed "s/\"/ /g" | awk '{print $5}')"
-        ${vWGET_CMD} ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/${vKEYRING_FILE} -O /tmp/${vKEYRING_FILE}
-        ${vWGET_CMD} ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/${vFEEDS_FILE} -O /tmp/${vFEEDS_FILE}
-        opkg install /tmp/${vKEYRING_FILE}
-        opkg install /tmp/${vFEEDS_FILE}
-        rm -fv /tmp/*${vNAME}*
+        if [ "$(opkg list-installed | grep "${vNAME}" | wc -l)" -lt "2" ]; then
+            vKEYRING_FILE="$(curl -s ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/ | grep ${vNAME}-keyring | sed "s/\"/ /g" | awk '{print $5}')"
+            vFEEDS_FILE="$(curl -s ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/ | grep ${vNAME}-packages-feeds | sed "s/\"/ /g" | awk '{print $5}')"
+            ${vWGET_CMD} ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/${vKEYRING_FILE} -O /tmp/${vKEYRING_FILE}
+            ${vWGET_CMD} ${vBASE_URL}/${vOWRT_VER}/packages/${vARCH}/packages/${vFEEDS_FILE} -O /tmp/${vFEEDS_FILE}
+            opkg install /tmp/${vKEYRING_FILE}
+            opkg install /tmp/${vFEEDS_FILE}
+            rm -fv /tmp/*${vNAME}*
+        else
+            echo " ###"
+            echo " ### ${vNAME}-packages repository already added ###"
+            echo " ###"
+        fi
+
         opkg update
+        
+        for vPKG in ${vALLOW_PKGS};do
+            opkg install ${vPKG}
+        done
+        
+        /etc/init.d/rpcd reload
+        /etc/init.d/cpu-perf disable
+        /etc/init.d/cpu-perf stop
+        /etc/init.d/internet-detector enable
+        /etc/init.d/internet-detector start
     fi
 fi
